modules/show/requirements.txt
redis
numpy
Pillow
plotly
opencv-python
psutil
requests
flask
PyTurboJPEG
